infer.fit() receives lnprobfn not lnprobfn_dict
When using jof.py to fit data, infer.fit() is called. There is a wrapper in the main function of jof.py around lnprobfn_dict called lnprobfn that is passed to infer.fit().  If the main function of infer.py is called, then lnprobfn_dict is defined at a higher scope so no error results. But when calling jof.py main function, lnprobfn_dict is not defined at the higher scope and an error results.

The fix is just to change the name of the lnprobfn passed to the Sampler.